+++ src/test/kotlin/lotto/domain/model/LottoNumberTest.kt
@@ -1,6 +1,5 @@
-package lotto.model
+package lotto.domain.model
 
-import lotto.domain.model.LottoNumber
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
